Fix more transforms and magnum renderer frames of reference.

+++ src/Rendering/MeshGenerator.cpp
@@ -34,7 +34,7 @@ void MeshGenerator::generatePoseMesh(vector<RigidBodyTransform> poses, vector<Ob
 
       Object3D& axes = parent->addChild<Object3D>();
       axes.translateLocal({static_cast<float>(translation.x()), static_cast<float>(translation.y()), static_cast<float>(translation.z())});
-      axes.scaleLocal({0.1 * scale, 0.1 * scale, 0.1 * scale});
+      axes.scaleLocal({(float) 0.1 * scale, (float) 0.1 * scale, (float) 0.1 * scale});
 
       objects.emplace_back(&axes);
       new RedCubeDrawable{axes, drawables, Magnum::Primitives::axis3D(), {(color * 123 % 255) / 255.0f, (color * 161 % 255) / 255.0f, (color * 113 % 255) / 255.0f}};
